third_party/WebKit/LayoutTests/fast/html/imports/import-element-removed-flag-expected.txt

This tests 'element removed flag' behavior defined in https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/imports/index.html#dfn-element-removed-flag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS staticImport.import is non-null.
PASS staticImport.import is null
PASS staticImport.import is non-null.
PASS dynamicImport.import is non-null.
PASS dynamicImport.import is null
PASS dynamicImport.import is non-null.
PASS dynamicImportEager.import is null
PASS dynamicImportEager.import is non-null.
PASS successfullyParsed is true

TEST COMPLETE
